Successfully use session cookie to auth user - Authenticate user before leaving quote - Adjust user in DB schema to use name and user id
