+++ api/helpers/quotesHelper.ts
@@ -17,7 +17,7 @@ export async function getUserQuotes() {
 export async function addQuote(
   quote: string,
   author: string,
-  authorId: number,
+  authorId: string,
   isUserQuote: boolean = true
 ) {
   await prisma.quote.create({
@@ -32,42 +32,42 @@ export async function addQuote(
   });
 }
 
-export async function deleteQuote(quoteId: number, apiKey: string) {
-  // Retrieve the quote to check the associated authorId
-  const quote = await prisma.quote.findUnique({
-    where: { id: quoteId },
-    select: { authorId: true }, // Select the authorId based on the quoteId
-  });
+// export async function deleteQuote(quoteId: number, apiKey: string) {
+//   // Retrieve the quote to check the associated authorId
+//   const quote = await prisma.quote.findUnique({
+//     where: { id: quoteId },
+//     select: { authorId: true }, // Select the authorId based on the quoteId
+//   });
 
-  if (!quote) {
-    throw new Error("Quote not found");
-  }
+//   if (!quote) {
+//     throw new Error("Quote not found");
+//   }
 
-  // Need to create userValidation function instead
-  // Retrieve the user to check the apiKey
-  const user = await prisma.user.findUnique({
-    where: { id: quote.authorId },
-    select: { apiKey: true },
-  });
+//   // Need to create userValidation function instead
+//   // Retrieve the user to check the apiKey
+//   const user = await prisma.user.findUnique({
+//     where: { id: quote.authorId },
+//     select: { apiKey: true },
+//   });
 
-  if (!user) {
-    throw new Error("User not found");
-  }
-  if (user.apiKey !== apiKey) {
-    throw new Error("Unauthorized: API key does not match");
-  }
+//   if (!user) {
+//     throw new Error("User not found");
+//   }
+//   if (user.apiKey !== apiKey) {
+//     throw new Error("Unauthorized: API key does not match");
+//   }
 
-  // Proceed with deletion if the apiKey matches
-  await prisma.quote.delete({
-    where: {
-      id: quoteId,
-    },
-  });
-}
+//   // Proceed with deletion if the apiKey matches
+//   await prisma.quote.delete({
+//     where: {
+//       id: quoteId,
+//     },
+//   });
+// }
 
-export async function loadQuotes(quotes: []) {
-  for (const { quote, author } of quotes) {
-    await addQuote(quote, author, 1, false); // Set isUserQuote to false for predefined quotes
-  }
-  console.log("Inspirational quotes loaded successfully!");
-}
+// export async function loadQuotes(quotes: []) {
+//   for (const { quote, author } of quotes) {
+//     await addQuote(quote, author, 1, false); // Set isUserQuote to false for predefined quotes
+//   }
+//   console.log("Inspirational quotes loaded successfully!");
+// }
