+++ api/index.ts
@@ -1,8 +1,7 @@
 import express from "express";
 import cors from "cors"; // Import the CORS middleware, which allows your server to handle cross-origin requests & server updates changes without reboot
-import { userRouter } from "./routes/user";
 import { quotesRouter } from "./routes/quotes";
-import { authRouter } from "./routes/authorisation";
+import { userRouter } from "./routes/user";
 
 const corsOptions = {
   origin: [
@@ -10,6 +9,7 @@ const corsOptions = {
     "https://random-quote-generator-api.vercel.app",
     "https://mindful-memos.peterforsyth.dev",
   ],
+  credentials: true,
 };
 
 const app = express();
@@ -19,7 +19,6 @@ app.use(cors(corsOptions)); // Apply CORS middleware with the specified options 
 
 app.use("/", quotesRouter);
 app.use("/", userRouter);
-app.use("/", authRouter);
 
 app.listen(8080, () => {
   console.log("Server is running.");
