+++ api/routes/user.ts
@@ -1,17 +1,74 @@
+import { WorkOS } from "@workos-inc/node";
 import { Router, Request, Response } from "express";
-import { newUser } from "../helpers/userHelper";
-import { User } from "@prisma/client";
+import cookieParser from "cookie-parser";
+import { handleUserLogin as saveUser } from "../helpers/userHelper";
+
+const workos = new WorkOS(process.env.WORKOS_API_KEY, {
+  clientId: process.env.WORKOS_CLIENT_ID,
+});
 
 export const userRouter = Router();
+userRouter.use(cookieParser());
 
-userRouter.post("/generateUser", async (req: Request, res: Response) => {
-  try {
-    const newUserData: User = await newUser();
-    res.status(201).json({
-      message: "User created successfully",
-      newUser: newUserData,
-    });
-  } catch (error) {
-    res.status(500).json({ error: "Failed to create user: backend" });
+userRouter.get("/login", (req, res) => {
+  const authorizationUrl = workos.userManagement.getAuthorizationUrl({
+    // Specify that we'd like AuthKit to handle the authentication flow
+    provider: "authkit",
+
+    // The callback endpoint that WorkOS will redirect to after a user authenticates
+    redirectUri: "http://localhost:8080/callback",
+    clientId: process.env.WORKOS_CLIENT_ID as string,
+  });
+
+  // Redirect the user to the AuthKit sign-in page
+  res.redirect(authorizationUrl);
+});
+
+// Triggered after /login redirects
+userRouter.get("/callback", async (req, res) => {
+  // The authorization code returned by AuthKit
+  const code = req.query.code as string;
+
+  if (!code) {
+    return res.status(400).send("No code provided");
   }
+
+  try {
+    const authenticateResponse =
+      await workos.userManagement.authenticateWithCode({
+        clientId: process.env.WORKOS_CLIENT_ID as string,
+        code,
+        session: {
+          sealSession: true,
+          cookiePassword: process.env.WORKOS_COOKIE_PASSWORD,
+        },
+      });
+
+    const { user, sealedSession } = authenticateResponse;
+    saveUser(`${user.firstName} ${user.lastName}`, user.id); // Saves user to database
+
+    // Store the session in a cookie
+    res.cookie("wos-session", sealedSession, {
+      path: "/",
+      httpOnly: true,
+      secure: true,
+      sameSite: "none",
+    });
+
+    res.redirect("http://localhost:5173");
+  } catch (error) {
+    return res.redirect("http://localhost:8080/login");
+  }
+});
+
+userRouter.get("/logout", async (req: Request, res: Response) => {
+  const session = workos.userManagement.loadSealedSession({
+    sessionData: req.cookies["wos-session"],
+    cookiePassword: process.env.WORKOS_COOKIE_PASSWORD as string,
+  });
+
+  const url = await session.getLogoutUrl();
+
+  res.clearCookie("wos-session");
+  res.redirect(url);
 });
